Add DLC guards to from_frame() methods and clean up endian.h
- Bug B16: Add DLC validation to 6 from_frame() decoders
  (SysThrottleSts, SteerDiag, BrakeDiag, SysDiagRpt, SesStatus,
  SebStatus) — all now return zero-initialized struct when DLC is
  insufficient, preventing garbage byte reads from latent callers
- Bug B17: Clean up read_be16 in endian.h — use conventional
  uint16_t OR pattern instead of unnecessary int16_t cast before
  shift

+++ shared/can/endian.h
@@ -24,7 +24,7 @@ inline int32_t read_be32(const uint8_t* buf) {
 }
 
 inline int16_t read_be16(const uint8_t* buf) {
-    return (int16_t(buf[0]) << 8) | int16_t(buf[1]);
+    return int16_t((uint16_t(buf[0]) << 8) | uint16_t(buf[1]));
 }
 
 } // namespace os
